pyproject.toml
[project]
name = "docker-gen"
version = "0.1.0"
description = ""
authors = [
    {name = "Wisnu Box",email = "hwisnu222@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "questionary (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
